+++ tests/r5-features.spec.ts
@@ -206,3 +206,30 @@ test.describe('r5-features.spec.ts', async () => {
   });
 
 });
+
+test.describe('R4 to R5', () => {
+  test.beforeEach(async ({page}) => {
+    await page.goto('/');
+    let mainPO = new MainPO(page);
+    await mainPO.acceptAllTermsOfUse();
+    const json = await PWUtils.readJSONFile('fixtures/local-storage-mock.R4.json');
+    await page.evaluate((mockQ) => {
+      window.localStorage.removeItem('state');
+      window.localStorage.setItem('fhirQuestionnaire', JSON.stringify(mockQ));
+    }, json);
+  });
+
+  test('should import R4 version from local storage', async ({page}) => {
+    await page.goto('/');
+    await page.getByLabel('Would you like to start from where you left off before?').click();
+    await page.getByRole('button', {name: 'Continue'}).click();
+    await page.getByRole('button', {name: 'Edit questions'}).click();
+    await expect(page.getByLabel('Data type', {exact: true})).toHaveValue(/coding/);
+
+    await expect(page.getByRole('radiogroup', {name: 'Create answer list'}).getByText('Yes')).toBeChecked();
+    const helpString = /^A plain text instruction/;
+    await expect(page.getByLabel('Help text', {exact: true})).toHaveValue(helpString);
+    const qJson = await PWUtils.getQuestionnaireJSONWithoutUI(page, 'R4');
+    expect(qJson.item[0].item[0].text).toMatch(helpString);
+  });
+});