+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 This project follows [Semantic Versioning](http://semver.org/).
 
 ## [10.0.6] 2025-05-14
+### Added
+- Added entry format extension.
 ### Fixed
 - Encoded ECL string of answerValueSet for SNOMED.
 - Fixed missing answer list layout for some item types.